pyproject.toml
[tool.poetry]
name = "qwen-lora-finetune"
version = "1.0.0"
description = "QLoRA / LoRA fine-tuning cho Causal LM (Qwen) trên GPU (RunPod / local)"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
transformers = ">=4.51.0"
datasets = ">=3.0.0"
peft = ">=0.14.0"
trl = ">=0.17.0"
bitsandbytes = ">=0.45.0"
accelerate = ">=1.6.0"
huggingface_hub = ">=0.24.0"
sentencepiece = "*"
wandb = ">=0.19.0"
scipy = "*"
python-dotenv = ">=1.0.0"
torch = ">=2.0.0"
# unsloth và flash-attn cài qua pip với variant CUDA/torch cụ thể (xem Dockerfile)
# unsloth = {git = "https://github.com/unslothai/unsloth.git", tag = "v2025.3.19"}

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
